patches from Vikas N Kumar <walburn@gmail.com> for making hunspell configurable.
This patch does the following:
1. configure.in
a) --disable-hunspell : option is given for user to disable Hunspell
explicitly
b) --with-hunspell-dictdir : option is given for user to select a
different Hunspell dictionary directory. Eg. /home/user/hunspell
c) No failure of configure if Hunspell is not found.

2. spellcheck-hun.c
a) Runtime checking of Hunspell *.aff and *.dic files in 4 default
directories, and 1 user supplied directory if given by configure.
b) The language "en" is mapped to both "en-US" and "en_US" files. The
original code has "en-US.aff" file for example, and if you download
the dictionary from Hunspell's website, then it has "en_US". The
current code supports both.
c) Only if the *.aff and *.dic file is not found is NULL returned. It
is pretty quick.

Signed-off-by: Vikas N Kumar <walburn@gmail.com> 


git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@28438 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6