Debugging patch from Amir P
Changes in this patches:
1. Added !test=comma-separated-list of features to enable/disable in the
code.
(Currently "altold" and "old_sane_morphism" can be enabled.)

2. Added !debug=comma-separated-list of function to which restrict the debug
output. Else it is printed from all the functions (if the verbosity level is
matched).

Both of these changes are implemented in error.c/error.h (so it is now for
errors and debug handling).
In principle debug.c/debug.h could be created for it (can do it later if you
think it worth the work).

If you issue the !var command, you will of course see the !debug and !test
variable setup. However, if they are non-null you will see prepended/appended
commas. This is an aesthetic problem that can be fixed later if needed.

3. Shorten lines in my code and improve some comments.

4. Aesthetic problem fix in the new sane_morphism(): skip checking morpheme
type combination if null_count>0, as it may be invalid due to morpheme
islands in alternatives (spurious warnings can be emitted otherwise).

5. Moved altlen() to utilities.c/utilities.h. (I also removed its copy at
read-dict.c, which is not included at this patch because my read-dict.c copy
currently contains my unfinished code for !!word).

6. Correct some misspelling in tokenize.c.

7. In command-line.c, added String to ParamType, for !test and !debug.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34021 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6