Patch, part 3 from Amir P.
This and the last two patches:


data/he/4.0.dict
- Add UNKNOWN-WORD.

link-grammar/build-disjuncts.c
(Not related to !!word: remove my old debug message.)

link-grammar/print.c
 Add print_sentence_word_alternatives(), used for !!word (in read-dict.c) and
for sentence debug (in tokeniez.c). If this is an inappropriate place for
this function I can move it to another place.

link-grammar/tokenize.c
-Remove the definition of print_sentence_word_alternatives().
- Change comments in my code.
- Use the new STRINGIFY() macro instead of local macro s_xstr().
- Change indentation and split long lines in my previous code.

link-grammar/dict-file/read-dict.c
- Fix typos in comments all over the file.
- Fix dict_order_wild() to not obey to a specified subscript when a wild-card
  is used.
- Show word splits if exist, also if the word itself is in the dictionary.
- Removed showing .I words (useful if there are tens of such words, like
  !!of). Can be re-enabled by !test=lookup-idioms in order to find if it is
useful anyway.
- The old code is still there (!test=altold, to be cleaned up soon). It uses
  the new dict_order_wild() so it has its new features.
- Add a comment to use UNKNOWN_WORD in order that wild-card will work (may
  need to change the wording).
- Spell guess not disabled (a 3 line code can be added for that).
- Enlarge field for umber of disjuncts from 5 to 6 (ru words may have >100000
  disjuncts).
- Prevent repeating showing information for the same tokens (tokens may
  repeat a lot for multiple splits of multiple tokens each). This is done by
recursively calling print_sentence_word_alternatives() to find out if the
token previously occurred.
- BTW, the split message "Word splits to:" repeats twice, but this can be
  considered a feature because the list can be long.

I used the word "String" because it can be a word, morpheme or a punctuation
(and idiom may be implemented too). Similarly, I used the word "Token"
because it can be a stem, a prefix, etc.
If desired, the token type can be decoded and referred by name.

Examples for Russian:
!!собравшихся
!!собра*

Note about a slight problem with words which are not stems (like
собравшихся):
Null suffixes are still shown for "=", even though they are not relevant in
this case.

Example for Hebrew:
!!והכלב
(cut & paste)

Examples for English:
!!do
!!doi*
!!doit
(It get split by the spell guesser)
!!of
(try !test=lookup-idioms)
!!.
(Not working - reason unknown)
!!"
(Showing nothing - this may or may not considered a bug.)
!!i.e.
!!i.e

link-grammar/link-includes.h
- Add Parse_Options to dict_display_word_info() and dict_display_word_expr().

link-grammar/print.h
- Add declaration for print_sentence_word_alternatives().

link-grammar/utilities.c
- Bug fix for !test and !debug: initialize with the empty string.
- Fix typos in comments all over the file.
- Fix left_print_string() for UTF8.

link-parser/command-line.c
- Minor format fix for printing the value of !debug and !test.
- Add Parse_Options argument to the invocation of dict_display_word_info()
  and dict_display_word_expr().



git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34090 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6