anysplit: random morphology tool from Amir
anysplit.c
- The Anysplit random word splitter.
- It includes all the code that can be concentrated in one file.

anysplit.h
Anysplit function prototypes.

autoconf-34586.patch
- Anysplit configure and Makefile support.
- --with-anysplit defines USE_ANYSPLIT .

data-any-34586.patch
- Definitions needed for Anysplit in the dict and affix files.

regex-n-more-34586.patch
- compile_regex(), match_regex(): Work directly on Regex_node.
- Improve the regex compile/match error messages.
- Dynamic allocation in sane_morphism().
- Always invoke sane_morphism().
- Affix file - reverse the word order to keep the original order.
- Boolean/TRUE/FALSE conversion.
- tokenize.c code function definition reorder.
- Move two alternatives functions to utilities.c.
- Typo corrections.
- Incorporate changes for Anysplit, under #ifdef USE_ANYSPLIT

anysplit-rest-34586.patch
- Added affix classes: REGPRE, REGMID, REGSUF, REGALTS, REGPARTS.
- Incorporate changes for Anysplit, under #ifdef USE_ANYSPLIT

---
Configuration:
The any/4.0.affix file contains the explanations on the added classes.
The maximum number of splits (REGPARTS) is set to 6, as needed for Tagalog. It causes Anysplit to
emit 1-6 candidate morphemes for each alternative. If REGPARTS is 0 or missing, Anysplit is
disabled.

The number of splits is set to 10 to 20, which can be far too much. It looks like it needs to be
reduced significantly.

The full word is always issued anyway. This skews the chances a little, but I was too lazy to fix
that and it may be unimportant.

The seed for rand_r() is currently constant. It means that for the same word, some partitioning
may be always missing. There is a commented-out line in the code to make the seed somewhat random
per word, so different random splits can be emitted for the same word in different sentences/runs.
It can be changed to use microseconds. The current behaviour is fine for tests that needs
repeatability. I can make it a configuration option (by affix class).

The regexes are currently set to ".*" . There can be multiple ones for each REGxxx, but currently
it is like defining "reg1|reg2" etc.  The check order of the morphemes is for maximum definition
flexibility and is documented. If needed, I can easily add negation (by ! before the regex) that
will work also in the 4.0.regex file, for a subset of the functionality of negative
lookahead/lookbehind.



git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34603 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6