From Amir: fixes for the window width ioctl
There is a problem in the terminal !width handling: The value of this
variable doesn't have any effect unless the process has no control
terminal.

The problem is that check_winsize() is always invoked, and it changes
the width back to the one taken from the control terminal. In order to
compare debug runs conveniently, I need to use a big width to prevent
folding, a thing that I could not do without stretching the window width
very much (including re-runs because I forgot to do so).

A good fix could be to use the !width value if it is set to anything
other than 0, else use the control terminal width (with a slight
drawback that can also be fixed, that !var would show 0 for !width
instead of the actual real terminal width.) I can do it if desired.

However, for now I introduced a simpler fix: use the actual terminal
width only if stdout is a terminal. So one can redirect to files using
the desired output width.


git-svn-id: svn+ssh://svn.abisource.com/svnroot/link-grammar/trunk@34924 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6