Fix printing the sentence words in verr_msg().
In the Wordgraph version, printing alternative[0] may print a total nonsense
morpheme sequences that may not even include the problematic words that caused
the reported problem (they could be in alternatives[1] and so on, this problem
is also in the pre-Wordgraph version). This is because in the Wordgraph version,
the "alternatives" array in the word-array don't really holds alternatives, but
it is a kind of a complex summary of them that may look rather strange.

For example for '50s,  ( [] means nothing at this place):

  word0: LEFT-WALL
      alt0: LEFT-WALL
  word1: '50s,
      alt0: ' s , ,   alt1: '50 50s s ∅   alt2: '50s 50 , []   alt3: [] ∅ [] []
  word5: RIGHT-WALL
      alt0: RIGHT-WALL

Fixed to print the tokenized sentence word/morphemes in the order in which they
appear in the sentence, without duplicates.