Remove word_array_pos. Fix typos.
The purpose of word_array_pos was to find the chosen disjuncts of a
Gword.  It also could serve as an easy way to find the all the subwords
at the position of a null word.

However, instead of that, the ordinal number of a Gword in this path can
be used as an index of chosen_disjuncts (now that the chosen disjuncts
array corresponds 1-to-1 to the Wordgraph path). Finding all the
subwords at the position of a null word is harder, but still possible,
using an algorithm similar to what sane_linkage_morphism() uses. In
principle, sane_linkage_morphism() can be extended to do that(using a
boolean arguments that tells it to work in that mode).