Fix the "dictcap" handling.
This handling is now ready for tests.
It is enabled by !test=dictcap.
In order to compare results by "diff" (actually "diff -i" for now, because the
capital letters are not restored yet), !test=dictcap,removeZZZ may be
used (to remove the dictcap tokens from the linkage results).

Dictcap handling of words that morpheme-split has not yet validated.