Handle long sequence of stripped off symbols as an unknown word
In case of a long sequence of symbols that get strip off a word by
left-strip or right-strip, a very large number of junk linkages
is created, like:
Found 4324 linkages ...
LEFT-WALL 92 Chapter[!] 10 ....v [...] ....y [........................]

To fix that, for now just accept the "word" as an unknown word.

A better fix (TODO) is maybe to strip real words off the sequence,
including cases when the symbols sequence is inside a "word".