Wchar path code got invoked when TCHAR==1
The result was a silent crash when the executable path contained even
number of characters, since the termination NUL was skipped (was in an
odd position in the path). On odd number of characters the result of
the conversion just got printed (as garbage) and ignored (and the
program worked fine).

This function should be rewritten. The search by specific DLL name is
not needed - only GetModuleHandle(NULL) should be used (btw, the DLL
name of x64 binaries, as set by the project file, is different -
linkgrammar-x64.dll). Also, the code uses a private implementation of
path_get_dirname(), when Windows already has PathRemoveFileSpec() that
does the same.