Check shorter prefixes even after a prefix has been found.
This is a bug fix.
Else משם is not getting split correctly, since מש is a prefix subword
which is longer than the prefix subword מ, and שם is a valid word.

Longer prefixes still have a priority over shorter ones when continuing
to search for the next prefix subword.  An example word in which this
occurs is משכבה, which all of its letters happen to be equal to prefix
subwords (כ is the next prefix subword that should be issued, not ש).