Add זה/זו (this m/f) and אלו (these) with m/f/s/p agreements.
The English "this.p" definitions were used.
It is not clear to me yet if the complex definition for these/those is
useful for Hebrew, hence a "this.p" style definition is used for אלו for
now.

The "this.d" English definition doesn't seem to me useful for Hebrew
at this stage, so I'm ignoring them for now.