Make sure the labels fully fit between adjacent words.
In the current linkage diagrams, if a link name is longer than the
space between centers of words, it is truncated. Moreover, if there are
h or d marks and this space is too small, the link name may even not
appear at all. In addition, the h and d marks don't appear if they
would cover the "+"'s in the diagram.

This change makes sure the adjacent words are spaced exactly enough to
make room for the link names + the h and d marks, and it solves
absolutely most of the cases in which a link name label is truncated in
the current batch examples.

For Hebrew, this fix currently affects many of the example sentence,
because the agreement links are relatively long and the words are
relatively short (and currently they are not even subscripted, a thing
that would make them longer but by itself would not solve this problem).

FIXME: Long link names between more distant words may still not
fit the space between these words. This is hopefully very rare.

Example of fixes:
1.
<                      +--------B*w-------+
<  +---Ds--+--Ss*q-+BIq+-Rn+--Sp-+----I---+
<  |       |       |   |   |     |        |
< the question.n is.v who we should.v invite.v
---
>                       +--------B*w-------+
>  +---Ds--+--Ss*q-+BIqd+-Rn+--Sp-+----I---+
>  |       |       |    |   |     |        |
> the question.n is.v  who we should.v invite.v

2.
< ----------Osm-
<  +----J---+
< -+    +VJl+-VJ
<  |    |   |
< on dogs.v , he
---
> -----------Osm
>  +-----J----+
> -+    +VJlgt+-
>  |    |     |
> on dogs.v   ,

3.
<     +------>T----->+
<     |      +<DVmvt<+--Xp--+
<     |      |       |      |
< LEFT-WALL ji  skaito.tevt .
---
>     +----->T----->+
>     |      +<DVmvt+--Xp--+
>     |      |      |      |
> LEFT-WALL ji skaito.tevt .

4.
< ---------------------+
< --------E-------+    |
< -----E-----+    |    |
<       +IDN+     |    |
<       |   |     |    |
< nndi два дня назад.e .
---
> ----------------------+
> -------E---------+    |
> ----E------+     |    |
>       +IDNE+     |    |
>       |    |     |    |
> nndi два  дня назад.e .

5.
<  +Ds+Ssc+
<  |  |   |
< ה= כלב רץ
---
<  +Ds+Sscm3+
<  |  |     |
< ה= כלב   רץ