Don't keep history while in batch mode.
Batch files are usually long and they overwrite the history of the
user's manual input. (A workaround could be to use a different directory
for running batches, but it not convenient.)

The !batch command that enters batch mode, if given as input, still gets
into the history file.

The history can still be pre-populated if desired, by reading files
when not in batch mode.