Don't try to split a lowercased spell-corrected word.
This happened, for example, when a sentence contained an unknown word in
a contraction, like "vicky's", and it got corrected to an uppercase
word, Vicki's in this case, when the corresponding lc vicki is an
unknown word. An assert() "NULL X-node" then occurred.

Note that this could happen only when the misspelled word was a
lowercase word that got fix to an uppercase word, since currently
uppercase words are not spell-corrected because they match a regex.