Add word ordinal number to subgraph display mode.
Subgraph display mode (requested by !test=wg:s) shows the unsplit words
as subgraphs, which the unsplit word as the subgraph title. This change
ads top this title the ordinal number of the unsplit word in the
tokenizing process.  This allows to match it to its node that is also
shown in the graph with the same ordinal number.