Change !spell to be the max. number of spell corrections per word.
As prior to this change, if the value of !spell is zero, no spell and
run-on corrections of unknown words are performed.

Else, use up to this many spell-guesses per unknown word.  This is an
API change. In that case, still (as before) the number of run-on
corrections (word split) of unknown words is not limited.