Fix crash in flattening complex wordgraphs (issue #146).
The problem was an incorrect order of handling wordgraph words, due to
dependency on the path position queue order, that was arbitrary but for
some reason worked for all the batch examples without an exception.  In
the problematic sentence, words that got issued too earlier got stack in
the path position queue until the RIGHT-WALL got issued, and then got
issued again. They got stuck because their next words were of the same
alternatives and could not get queued until the RIGHT-WALL has
encountered.

Due to that, the number of issued words was greater than pre-calculated,
triggering an assert().

The fix was to insert word (wordgraph nodes) into the path position
queue according to reverse order of their depth level in the wordgraph
hierarchy.  The rational is that in order to flatten the wordgraph,
there is a need to scan the wordgraph starting from the deepest
alternatives so words at higher level alternatives will never get issued
before words in lower level ones.

Also in this patch:
Enhance the debug printings to also print the gword node_num.
Add asserts to abort if a word is issued after the RIGHT_WALL,
or if a word is issued more than once.