Regex file: Implement negative regex'es.
Add the following syntax:
REGEX_NAME: !/regex/

Regex'es that are so preceded by !, if they match a token, stop
further match tries of the same regex name. Thus, they can serve
as a kind of a negative look-ahead.

Introduce an EMOTICON negative match !/^"/  to prevent matching
quoted strings.

Add an example sentence that used to be parsed as correct because
of an misidentified emoticon.

Add a changelog line.