In case sentence_split() fails, make sure an error is printed.
An error is "expected" to be printed anyway, but in case of an internal
error nothing can be guaranteed. Because we don't have yet an error API,
we don't know if an error has already been generated at a lower level.
So the fix here is to always print an "Internal error" message.

The case that caused this fix was a bug in separate_word():
A regex match was not done for a word that is not in the dict, and the
word remained unknown. Later, sentence_in_dictionary(), that is supposed
to print an error if there are unknown words, missed this problem
because it explicitly check each word using find_word_in_dict(). So no
error got printed.
A#