Fix wordgraph flattening: check hierarchy depth.
Advancing in the wordgraph always in the same alternative level was
wrong. Actually it is needed to advance in the same alternative depth
level (heir_depth) or a higher one. Else higher levels could be skipped
when their turn arrives, and visited only after the RIGHT-WORD is
encountered (which is clearly wrong).

Add wordgraph_hier_position() on all the wordgraph words, in order to
use it letter for advancing in the wordgraph.

This token caused a mess due to that bug:
([<vc-do]>

Its structure is (|, + and - are ascii-art constructs of the graph):

LW - ([<vc-do]> - RW
       |
       +- ([<vc-do] - >
       |
       +- ( - [ - < - vc-do>
                        |
                        +- vc-do - >

The next token to be scanned after < should be vc-do . Without the fix,
the next one is RW.