Fix a "NULL X-node" assert due to incorrect stem-suffix split
Background :
When a word is unknown and spelling is enabled, a spell guess is
performed on it, and its result words are checked for validity.  A
result word is valid if it is in the dict file, or if it can split.

To that end, suffix_split() is supposed to return a Boolean, saying if
the word can split. It is invoked with its unsplit_word argument as
NULL, so it doesn't actually issue the split as an alternative.

There were two problems in this mechanism:
1. add_alternative_with_subscr(), which return status is needed for
that purpose, was not called (because the check is done, as described
above, using NULL as unsplit_word).
2. Even if this is fixed, add_alternative_with_subscr() retuned true
unconditionally.

Consider the case of an input word (de) ginge.  It is currently
not in the dict. However, =e is the dict, and also ging.  So
add_alternative_with_subscr() was called to add "ging =e" as
an alternative. However, after (1) above has been fixed, it (correctly)
found that ging is not a stem according to the definition of STENSUBSCR.
So nothing got issued also when it got invoked not for spell validity
checking. But due to problem (2), a true value was returned, wrongly
indicating that it is valid. Later in determine_word_expressions()
it was not fond in the dict as assumed, resulting in a NULL X-node which
triggered an assert().

Also add in the code comments that explain this check.