Python binding: Add Sentence.num_valid_linkages()
Currently finding the number of linkages is problematic, as
the Sentence.parse() returns an iterator. There is a need to first
generate all the linkages into a list, and then use len() on it.

Just adding a binding of sentence_num_valid_linkages() is not good
enough, because currently the parsing is not actually done by
Sentence.parse() until the first linkage is obtained from the iterator.

To solve that, this fix implements the iteration using a class,
when the class initializer performs the parsing so num_valid_linkages()
returns a valid result even before obtaining the first linkage.