Fix handling of stems with non-LL links
When morphology is hidden (the default) there is a sever bug regarding
printing of stems with non-LL links which causes nonsense printings and
crashes.

The fix is to handle hidden morphology of stems just like any other
combined morphemes, and remove the morphology links (currently LL*)
in a separate step.

The links are so identified by the function is_morphology_link().
Currently it checks for LL links, but it can be extended to
check for a configurable list of links.

If a word remains with no links due to this removal, it is removed from
the linkage altogether.

The results is that currently the stem-suffix links remain in the printout,
just like the prefix-word links currently behave:

With HIDE_MORPHO:
                            +---S3s---+
    +----W---+---EA--+---A--+  +-VPast+
    |        |       |      |  |      |
LEFT-WALL çok.ad şişman.a adam  geldi.past3s

Without:
                            +-----S3s-----+
    +----W---+---EA--+---A--+    +--VPast-+
    |        |       |      |    |        |
LEFT-WALL çok.ad şişman.a adam gel.= =di.past3s

(Note VPast, previously it was not printed, and if words got added to
the sentence, it got printed in a totally incorrect place.)