tr: Restore transitive verb, rearrange connector order
The original connector order prevented the linkage of ve.j-v.
Now all the sentence get parsed, when at least one of the parses seem
correct (not totally sure - I don't know Turkish...).

Also, a change is introduced in the singular nouns as an example for
enforcing S+ or O+, but never both at once (as happens without this
fix).

"adam geldi" still has incorrect parses when O+ is used instead of S+
(it also seem to have the correct parse with S+).
It looks like S+ should be enforced by using it together with W- in a
separate disjunct. But then ve.j-n related stuff should be fixed too,
and I leave that to the original author to continue.