Python bindings: Fix the new bindings of spell_guess.
The patch "Change !spell to be the max. number of spell corrections per
word." introduced an API change.
My first fix of the corresponding Python bindings is buggy.
This fix tries to make it better.

The behavior after this fix is this:

If the spell_guess option is set to 0 or a positive integer, the max
number of corrections is set to this number.
A value of False is like specifying 0.
A value of True is like specifying 7.
Any other value is an error (this already has a test).

Also:
- Document it.
- Add a test of this behavior.