sat-solver: Don't terminate on a disconnected linkage
Instead, just continue, as the original code did (but without
producing the linkage).

Terminating on a disconnected linkage just prevents producing the rest
of valid linkages, including not producing linkages for a sentence with
valid linkages like that: This is the xxxing problem.

For testing, it is possible to see the disconnected linkages too, using:
!test=linkage-disconnected

Additional things that get fixed by this patch:
- Redundant linkage creation (and then discarding it) for disconnected
  linkages.
- Memory leak due to discarding the linkage before freeing the
connectors/disjuncts.