sat-solver: Fix postprocessing handling and "bad" linkages handling.
1. Add setting N_violations and pp_violation_msg.

2. Remove the following comment in sentence_num_violations():
/* The sat solver (currently) fails to fill in link_info */

3. In link-parser, check sentence_num_violations() also after
linkage_create(), since only then N_violations is set by the sat solver.

4. Since now "bad" linkages are not printed by default, don't print
the Linkage PP OK/ Linkage PP NOT OK messages.

5. We cannot print the sat solver stats unconditionally any more,
since "bad" linkages are not usually printed. We can also cannot check
if the "bad" options is set, and don't have API to print linkage stats.
So print the sat solver stats only on verbosity > 1.
(Now verbosity is also used for debug level - maybe there should be a
separate debuglevel.)
This is also needed in order to prevent stat printouts on verbosity==0.

Comment: The first (optional) step of the preprocessing is not done,
and hence is_short is set to false.