sat-solver: Fix memory leak due to empty-words.
The empty-words disjuncts and connectors are created in
get_next_linkage() only for the purpose of using remove_empty_words()
later (which is needed in order to renumber the words). However,
discarding the empty-words prevents freeing their disjuncts and
connectors.

This fix prevent the memory leak by implementing these fixes:
1. Concatenate the empty-words disjuncts to the disjunct of the first
sentence word(that must exists in case a a sentence with an empty word
has a linkage).
2. Don't allocate memory for the connectors. Instead - use a dummy
connector (the connectors are referenced in some other parts of the
code, so random or NULL value causes a crash).