sat-solver: Always try to find the first linkage.
Also (see below):
- Use sent->lnkages as cache.
- Validate the linkage index.

The rational for always trying to find the first linkage:
1. For an interactive user, finding the first linkage is anyway done
unconditionally by link-parser.
2. For a batch run, it is useless for the solver to always return
success. There is a need to know if there is at least one valid linkage.
3. For the library user there is no use of the sat solver without
finding at least the first linkage.

If a linkage is found, it is cached in sent->linkages, so when it is
requested it is not computed again. This mechanism allows caching of
further linkages. The next linkage is requested from the SAT solver if
the requested linkage (by index) is not cached. For such linkages the
index is now validated to be the next linkage number.