sat-solver: Fix a crash after a panic.
This patch fixes 2 problems:
1. If the selected parser is the SAT solver, in panic mode the standard
parser still gets used, a thing that causes a memory management mess.
This mess is not fixed in this patch. Instead, in that case the SAT
solver is used also in panic mode. This is currently not too useful
since parsing with null words is not yet supported there. It is left to
the user to disable panic mode in that case.

2. A crash that happens if the SAT solver is used in panic mode.