sat-solver: Remove unused code (speedup).
There is an observable performance increase when removing this code.
I could not find the reason for this code even in the initial check-in
of the SAT solver code.
I produced detailed linkages of en/4.0.batch and en/4.0.fix-long.batch
(using Minisat-2.2, because the current Minisat library has a problem
there) before and after this change, and there indeed were no differences.
So it seems pretty safe to not include this code.