append_string(): Fix two bugs
For a background, see:
https://securityblog.redhat.com/2014/03/12/the-trouble-with-snprintf/

1. A possibility to return a negative value (on error) is documented in
the manual. If from some reason an error occurs, a fatal malfunction will
result.  Check for a possible error.

2. If the vsnprintf() buffer is insufficient, the print result gets
silently truncated to the buffer size (with a proper NUL termination).
This may be considered a problem, as the program output is then
intermixed in the middle of a line with further output from other
sections of the program.
In addition (from the manual): "If  the  output was truncated  due to
this  limit, then the return value is the number of characters
(excluding the terminating null byte) which would have been written to
the final string if enough space had been available."
But the buffer is currently extended by this returned value. Hence
output from subsequent append_string() calls will be hidden due to the
resulted NUL in the middle of the buffer. This insufficient buffer
problem is not fatal for ISO C99 vsnprintf() (as currently in GNU C),
which doesn't return -1 on an insufficient buffer (else it is, see (1)
above).

In the same occasion, simplify the function somewhat.