match_table: Make the entries to be per uc connector parts
Currently connectors with different uppercase parts can be concatenated
to the same entry, so form_match_list() may return connectors which have
no chance to match due to their mismatched uppercase portion.

This change ensures that each match table entry contains only
connectors with identical uc parts.

In order to to check if the uppercase part is the same,
calculate_connector_hash() is modified to calculate and cache in the
Connector_struct additional values: the start of the uc and lc parts,
and the uc part length. These will also be used later in to speed up
connector comparisons.