Store the String_set struct address for debug verifications
Issue #206.
In order to save CPU time and memory references on string comparisons,
when it is known that both strings are in the same string-set, it is
possible exploit the fact that the memory address of these strings is
equal iff the strings themselves are equal.

However, uncareful use of this method may lead to strange bugs if the
strings are not from the same string-set (e.g. if one of them is a
constant string, or they are from different string-sets). Hopefully,
validating that the strings are appropriate for that method may easy
this problem.

This change stores - in DEBUG mode only - the String_set struct address
after the ending NUL byte of the string. The function that verifies the
strings (to be introduced in the next commit) compares these addresses.
This is not a full-proof verification, but the chances for false
positives are very low.