Use string_set_cmp() to compare connectors.
This doesn't cause a big speedup, but it is a start.
More comparisons may use this method.

(Using it for disjunct strings is not yet safe, because currently some
of them are from the Dictionary string set and some from the Sentence
string-set. Similarly, using it in the PP stage is not safe too.).