Don't set connector length-limit for too short sentences.
If the sentence length is not greater than opt->short_length, this only
wastes time. This saves CPU in the order of 1% on en/corpus-fixes.batch.

The change is done only in the regular parser. For the SAT one there
is no loop that can be eliminated.

Several things to note:
1. The ZZZ connector length-limit could be set to 1, but the loop setup
time for too short sentences is bigger than the saved parsing time.
For SAT (no setup loop) this will help.

2. If desired, the ZZZ connector length-limit could be set to 1 here for
long enough sentences, in which the loop is still done anyway.

3. The loop in set_connector_length_limits() can be eliminated
altogether with if more preparations are at the dictionary processing
step.