Test suite: Add a basic test of the SAT parser
In case the library has been configured without the SAT parser, the test
suite skips the test with a reason message (that can be seen when the
tests are run with the -v argument).

For the basic testing of the SAT parser, the first parse it currently
happens to emit for "this is a test" is used, along with its
constituents printing output.