Add connector match caching
Due to the way in which the match-lists are retrieved (by connector uc
part) the match list tend to have a long series of disjuncts with the
same connectors - left connector for ml and right for mr.  When the
connectors are not the same, they have the same uc part.

This change tries to utilize these features for fast connector matching,
in these ways:
1. If the cache contains the match status of the connector from the
previous match list element - use it. This is a string-set comparison,
but without using the function string_set_cmp() because c_con->string
(the cached connector) may be NULL for the first list element.
(string_set_cmp() is used in other places of connector comparisons,
which is hopefully enough for debug purposes.)

2. Check if any of the connectors doesn't have a lc part. If so, they
match because we know all the connectors in the match list have the
same uc part like the checked connector.

Benchmarks comparing to current master (the improvement is mainly for long
sentences):
avg 100 runs of best of 2:
~1% en/corpus-basic.batch
~0% ru/corpus-basic.batch
~8% en/corpus-fixes.batch

avg 10 runs of best of 2:
~12% en/corpus-fix-long.batch [without last 3 sentences]