Don't accept a contracted word which is not in the dict as spell correction
The following garbled word:
test,'s,
gets the spell correction:
Estes's

It can be split so it is currently considered as a valid word. However,
this is wrorg, as Estes is actually not in the dictionary.

Add a check to validate, in a spell result test, that the base word is
actually in the dictionary.