Fix handling of a linkage null-word prefix
The parse result of the sentence: *w/milk
is displayed as: [w] milk
instead of: [w/] milk

The problem is in displaying prefixes/suffixes: the function
add_morpheme_unmarked() supposes that they include an INFIX_MARK_L
(length 1) INFIX_MARK character. This is incorrect in case INFIX_MARK
is not defined.