Add a check for the match list integrity
Recursive calls of do_count() and mk_parse_set() invoke form_match_list()
again while the parent functions still traverse the match lists of
previous form_match_list() invocations. Havoc would happen if the later
form_match_list() invocations would change the match_left/match_right
fields of disjuncts which has not been traversed yet.

Such a problem has never been observed. However, currently I don't have
a proof it can never happen. This change adds an integration validity of
the match lists by storing a match list ordinal number in each disjunct,
and validating that for each disjunct in the list it always remains the
same.

The check is enabled for now in production versions, but the intention
is that it will be later done only in debug mode. To that end,
uncommenting the "#ifdef DEBUG" in structures.h.