Fix efficiency bug in handling Russian (redundant regex match)
In case of words that can split, a regex check is tried anyway.
This doesn't change the results (as verified by a detailed run of
ru/corpus-basic.batch), but is a major efficiency hit due to the large
ru/4.0.regex file.

A benchmark (7 runs of best-of-7) shows a speed up of 14%.