Remove skipping the length_limit setting for short sentences
This is needed so the just-introduced setup of ZZZ length-limit=1
will work.

FIXME: Setting the length_limit for short sentences (<short-length)
which don't contain empty-words takes observable CPU time.
This can be fixed be adding a flag to indicate that the sentence
contains an empty word. But it may be better to just use another method
for length_limit (like using connector-descriptor pointers in the
connectors).