Fast matcher: Fix length-limit check of the match-list connectors
A match-list contains only connectors with the same UC prefix.  However,
some connectors may have the same UC prefix but a different
length_limit, e.g. X and Xc which have unlimited length, and Xd which
have short_length.

If lc or rc would be an unlimited-length connector (e.g. X)
and the match-list would contain matching connectors with different
length-limit, e.g. Xc and Xd, the whole match-list could be missed
if the first connector is a short_length one. (If lc or rc would be
a short_length connector then we don't care.)

Such a thing has not been observed by a debug print. In addition,
absolutely no linkage in all the languages got changed due to this fix
(in English the fixes.batch has been verified too).

However, the code seems wrong in principle and in order to be on the safe
side it is fixed here. (This is a very subtle issue.)