Bug fix in the test_enabled macro (issue #255)
The intention was to test the first character of this variable.
The variable itself is never NULL.

This problem caused the following on OS X El Capitan:
sat-encoder.cpp:1320:38: error: comparison between pointer and integer ('int' and 'const char *')
      display_linkage_disconnected = test_enabled("linkage-disconnected");
                                     ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
../../link-grammar/error.h:57:9: note: expanded from macro 'test_enabled'
        (('\0' != test) ? feature_enabled(test, feature) : NULL)

It was also an inefficiency bug: The feature_enabled() function was
called always instead of only when a test is set.